finalized collate part
