+++ tests/leap_c/utils/test_collate.py
@@ -7,7 +7,7 @@ from acados_template.acados_ocp_iterate import AcadosOcpFlattenedBatchIterate
 
 from leap_c.diff_mpc.data import AcadosOcpSolverInput
 from leap_c.diff_mpc.function import AcadosDiffMpcCtx
-from leap_c.utils import ACADOS_DIFF_MPC_COLLATE_FN_MAP, collate_acados_diff_mpc_ctx, collate_torch
+from leap_c.utils import collate_acados_diff_mpc_ctx, collate_torch
 
 
 def _make_ctx(status: int = 0, x0: np.ndarray | None = None) -> AcadosDiffMpcCtx:
@@ -39,7 +39,7 @@ def ctx_batch() -> list[AcadosDiffMpcCtx]:
     ]
 
 
-def test_collate_acados_diff_mpc_ctx(ctx_batch: list[AcadosDiffMpcCtx]) -> None:
+def test_collate_acados_diff_mpc_ctx(ctx_batch: list[AcadosDiffMpcCtx]):
     """Direct stacking of contexts produces a batched context."""
     result = collate_acados_diff_mpc_ctx(ctx_batch)
 
@@ -50,13 +50,7 @@ def test_collate_acados_diff_mpc_ctx(ctx_batch: list[AcadosDiffMpcCtx]) -> None:
     assert result.log is None
 
 
-def test_acados_diff_mpc_collate_fn_map_keys() -> None:
-    """The map exposes the ctx type as its sole key."""
-    assert AcadosDiffMpcCtx in ACADOS_DIFF_MPC_COLLATE_FN_MAP
-    assert ACADOS_DIFF_MPC_COLLATE_FN_MAP[AcadosDiffMpcCtx] is collate_acados_diff_mpc_ctx
-
-
-def test_collate_torch_with_ctx(ctx_batch: list[AcadosDiffMpcCtx]) -> None:
+def test_collate_torch_with_ctx(ctx_batch: list[AcadosDiffMpcCtx]):
     """collate_torch handles dicts containing both tensors and ctx objects."""
     samples = [
         {"x0": torch.tensor([1.0, 2.0]), "ctx": ctx_batch[0]},
@@ -74,7 +68,7 @@ def test_collate_torch_with_ctx(ctx_batch: list[AcadosDiffMpcCtx]) -> None:
     assert batch["ctx"].status.tolist() == [0, 1]
 
 
-def test_collate_torch_plain_tensors() -> None:
+def test_collate_torch_plain_tensors():
     """collate_torch still handles plain tensors without ctx objects."""
     samples = [torch.tensor([1.0]), torch.tensor([2.0])]
     batch = collate_torch(samples)
